Fix mariadb binding on ubuntu servers.
with the newer ubuntu versions the bind address is in 5-server.cnf instead of my.cnf while my.cnf still exists. This if-else branch never accesses the code to fix 50-server.cnf if my.cnf still exists so moving the 50-server.cnf to the top fixes the issue.

tested on multiple 16.04 and 18.04 vpsses.